improve the fetchConfigDesc func
